+++ client/src/assets/styles/components/BookShowcase/bookshowcase.scss
@@ -1,0 +1,50 @@
+@use "../../../SASS/" as *;
+
+body {
+    div#root {
+        section.showcase-section {
+            background-color: $background-light;
+            border-radius: 50px;
+            width: 80%;
+            margin: auto;
+            padding: 50px;
+
+            h2.showcase-title {
+                font-family: $logo-font;
+                font-size: 50px;
+                color: $fontcolor;
+                text-align: center;
+                margin-bottom: 50px;
+            }
+            
+            div.showcase-wrapper {
+                display: flex;
+
+                div.profile-book {
+                    width: 25%;
+                    display: flex;
+                    justify-content: center;
+                    align-items: center;
+                    flex-direction: column;
+        
+                    h3.profile-book-title {
+                        color: $fontcolor;
+                        font-family: $logo-font;
+                    }
+                }
+        
+                div.profile-book:hover {
+                    transform: scale(1.1);
+        
+                    img.profile-book-image {
+                        filter: drop-shadow(1px 1px 5px $main);
+                    }
+        
+                    h3.profile-book-title {
+                        color: $main;
+                    }
+                }
+            }
+        }
+    }
+}
